moving data methods into mixin
